GJH - 사원관리 페이지 완료

+++ src/main/java/com/project/hot/employee/model/dto/RequestEmployee.java
@@ -21,7 +21,7 @@ public class RequestEmployee {
 	private String employeePassword;
 	private String employeeId;
 	private String employeeAddress;
-	private int employeeSalary;	
+	private long employeeSalary;
 	private Date employeeHireDate;
 	private Date employeeResignationDay;
 	private int employeeTotalVacation;
